+++ frontend/src/components/AddPermissions/ForestClientSection.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Button from "@/components/Button.vue";
+import Button from "@/components/UI/Button.vue";
 import SearchLocateIcon from "@carbon/icons-vue/es/search--locate/16";
 import ForestClientCard from "@/components/AddPermissions/ForestClientCard.vue";
 import { AppActlApiService } from "@/services/ApiServiceFactory";
@@ -9,7 +9,7 @@ import { type FamForestClientSchema } from "fam-app-acsctl-api";
 import InputText from "primevue/inputtext";
 import { ErrorMessage, Field } from "vee-validate";
 import { computed, h, ref, watch } from "vue";
-import NotificationMessage from "../NotificationMessage.vue";
+import NotificationMessage from "../UI/NotificationMessage.vue";
 import type { ForestClientNotificationType } from "@/types/NotificationTypes";
 
 const props = withDefaults(
