+++ frontend/src/views/AddAppPermission/index.vue
@@ -44,8 +44,8 @@ import UserNameInput from "@/components/AddPermissions/UserNameSection.vue";
 import RoleSelectTable from "@/components/AddPermissions/RoleSelectTable.vue";
 import type { FamRoleDto } from "fam-admin-mgmt-api/model";
 import ForestClientSection from "@/components/AddPermissions/ForestClientSection.vue";
-import BoolCheckbox from "@/components/BoolCheckbox.vue";
-import Button from "@/components/Button.vue";
+import BoolCheckbox from "@/components/UI/BoolCheckbox.vue";
+import Button from "@/components/UI/Button.vue";
 import type { FamAccessControlPrivilegeCreateRequest } from "fam-admin-mgmt-api/model";
 import ConfirmDialog from "primevue/confirmdialog";
 import { useConfirm } from "primevue/useconfirm";
